refactor: extract code editor module
