+++ desktop/src/editor/mod.rs
@@ -1,0 +1,4 @@
+pub mod code_editor;
+
+pub use code_editor::CodeEditor;
+pub use code_editor::THEME_SET;
