+++ desktop/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod app;
 pub mod components;
 pub mod modal;
+pub mod editor;
